+++ src/App.css
@@ -1,136 +1,4 @@
-@media screen and (max-width: 480px) {
-  .desktopHome {
-    display: none;
-  }
-
-  .mobileHome {
-    position: absolute;
-  }
-
-  nav {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    justify-content: space-around;
-    padding: 40px 40px;
-    max-width: -70px;
-    max-height: 496px;
-    position: absolute;
-    margin-left: -370px;
-    top: 30px;
-    padding-left: 5px;
-    background: rgba(255, 255, 255, 0.04);
-    backdrop-filter: blur(40.774227142333984px);
-  }
-
-  .horizonal-line {
-    display: none;
-  }
-
-  .logo {
-    margin-left: -700px;
-    background-color: white;
-    border-radius: 50px 50px;
-    position: absolute;
-    top: 30px;
-  }
-}
-
-@media screen and (min-width: 481px) and (max-width: 768px) {
-  .navBar {
-    display: none;
-  }
-}
-
-@media screen and (min-width: 820px) {
-  .nav {
-    display: none;
-  }
-
-  .routeMobile {
-    display: none;
-  }
-
-  .navBar {
-    margin-left: 250px;
-  }
-}
-
-nav {
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-around;
-  padding: 40px 40px;
-  max-width: 830px;
-  max-height: 96px;
-  position: absolute;
-  margin-left: 700px;
-  top: 30px;
-  padding-left: 5px;
-  background: rgba(255, 255, 255, 0.04);
-  backdrop-filter: blur(40.774227142333984px);
-}
-
-nav a {
-  list-style: none;
-  color: white;
-  text-transform: capitalize;
-  font-family: Barlow Condensed;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-  letter-spacing: 2.7px;
-  text-decoration: none;
-  padding-top: 20px;
-  padding-left: 40px;
-}
-
-nav a:hover {
-  color: #d0d6f9;
-  transition: all 100ms ease-out;
-}
-
-nav a:hover::after {
-  content: "";
-  width: 10%;
-  height: 2px;
-  background-color: white;
-  position: absolute;
-  bottom: -2px;
-  left: 50px;
-}
-
-nav a.active::after {
-  content: "";
-  width: 10%;
-  height: 2px;
-  background-color: #d0d6f9;
-  position: absolute;
-  bottom: -2px;
-  left: 50px;
-}
-
-.navBar li {
-  color: white;
-  display: none;
-}
-
-.logo {
-  margin-left: -1750px;
-  background-color: white;
-  border-radius: 50px 50px;
-  position: absolute;
-  top: 30px;
-}
-
-.horizonal-line {
-  position: absolute;
-  margin-left: -1060px;
-}
-
-#mobile-view,
-#times {
-  color: white;
-  display: none;
+header {
+    width: fit-content;
+    overflow: hidden;
 }
