+++ src/Pages/styles/navbar.css
@@ -1,0 +1,131 @@
+.navBar {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    width: 98%;
+    overflow: hidden;
+    margin: 40px 0 0 40px;
+    position: absolute;
+    z-index: 1;
+}
+
+.logo {
+    width: 3.5%;
+    height: 50px;
+    border-radius: 50px;
+    background-color: white;
+    cursor: pointer;
+}
+
+hr {
+    color: gray;
+    width: 50%;
+    height: 3px;
+    margin: 0 0 0 1%;
+}
+
+nav {
+    display: flex;
+    align-items: flex-end;
+    justify-content: space-around;
+    padding: 30px 0;
+    margin: 0 5% 0 0;
+    width: 50%;
+    height: fit-content;
+    background: rgba(255, 255, 255, 0.04);
+    backdrop-filter: blur(40.774227142333984px);
+}
+
+nav a {
+    color: white;
+    text-transform: capitalize;
+    font-family: Barlow Condensed;
+    font-size: 12px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal;
+    letter-spacing: 2px;
+    text-decoration: none;
+    padding: 0;
+    cursor: pointer;
+}
+
+nav a:hover {
+    color: #c8cff9;
+}
+
+nav a.active::after {
+    content: "";
+    width: 10%;
+    height: 2px;
+    background-color: #d0d6f9;
+    position: absolute;
+    bottom: -2px;
+    left: 50px; /* Adjusted the left property */
+}
+
+#mobile-view {
+    display: none;
+}
+
+@media screen and (max-width: 1024px) {
+    hr {
+        width: 35%;
+    }
+    .logo{
+        width: 50px;
+    }
+}
+
+@media screen and (max-width: 768px) {
+    hr {
+      width: 10%;
+    }
+
+    .navBar {
+        gap: 1%;
+        width: 92%;
+    }
+
+    nav { 
+        width: 100%;
+        margin: 0;
+    }
+}
+  
+@media screen and (max-width: 426px) {
+    hr {
+        display: none;
+    }
+
+    nav {
+        display: none;
+    }
+
+    .navBar {
+        margin: 40px 0 0 20px;
+    }
+
+    #mobile-view {
+        display: flex;
+        margin: 0 0 0 70%;
+        color: white;
+    }
+
+    #times {
+        cursor: pointer;
+    }
+
+}
+
+@media screen and (max-width: 320px) {
+
+    #mobile-view {
+        margin: 0 0 0 65%;
+    }
+
+}
+  
+
+  
+
